# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/health": {
      "get": {
        "operationId": "HealthController_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "GUEST",
        "tags": [
          "Health"
        ]
      }
    },
    "/api/user/getUsers": {
      "get": {
        "operationId": "UserController_getUsers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page Number",
            "schema": {
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Order direction: ascending  (asc) veya descending (desc)",
            "schema": {
              "example": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersResponseDataDto"
                }
              }
            }
          }
        },
        "summary": "ADMIN",
        "tags": [
          "User"
        ]
      }
    },
    "/api/user/updateUser/{userId}": {
      "patch": {
        "operationId": "UserController_updateUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "USER",
        "tags": [
          "User"
        ]
      }
    },
    "/api/user/deleteUser/{userId}": {
      "delete": {
        "operationId": "UserController_deleteUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "USER - ADMIN",
        "tags": [
          "User"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "GUEST",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "GUEST",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/logout": {
      "get": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponseDto"
                }
              }
            }
          }
        },
        "summary": "USER",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "USER",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/post/getPosts": {
      "get": {
        "operationId": "PostController_getAllPosts",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page Number",
            "schema": {
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Order direction: ascending  (asc) veya descending (desc)",
            "schema": {
              "example": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "GUEST",
        "tags": [
          "Post"
        ]
      }
    },
    "/api/post/getPost/{slug}": {
      "get": {
        "operationId": "PostController_getPostBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "GUEST",
        "tags": [
          "Post"
        ]
      }
    },
    "/api/post/getSlugs": {
      "get": {
        "operationId": "PostController_getSlugs",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "GUEST",
        "tags": [
          "Post"
        ]
      }
    },
    "/api/post/createPost": {
      "post": {
        "operationId": "PostController_createPost",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "ADMIN",
        "tags": [
          "Post"
        ]
      }
    },
    "/api/post/updatePost/{postId}": {
      "patch": {
        "operationId": "PostController_updatePost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "ADMIN",
        "tags": [
          "Post"
        ]
      }
    },
    "/api/post/deletePost/{postId}": {
      "delete": {
        "operationId": "PostController_deletePost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "ADMIN",
        "tags": [
          "Post"
        ]
      }
    }
  },
  "info": {
    "title": "Rental Car API",
    "description": "API documentation for the Rental Car system",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "dateOfBirth": {
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "isAdmin": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "_id",
          "name",
          "surname",
          "dateOfBirth",
          "email",
          "isAdmin",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "perPage": {
            "type": "number",
            "example": 10
          },
          "totalPages": {
            "type": "number",
            "example": 1
          },
          "currentPage": {
            "type": "number",
            "example": 1
          },
          "pageStartIndex": {
            "type": "number",
            "example": 1
          },
          "hasPrev": {
            "type": "boolean",
            "example": false
          },
          "hasNext": {
            "type": "boolean",
            "example": false
          },
          "prev": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "next": {
            "type": "object",
            "example": null,
            "nullable": true
          }
        },
        "required": [
          "perPage",
          "totalPages",
          "currentPage",
          "pageStartIndex",
          "hasPrev",
          "hasNext",
          "prev",
          "next"
        ]
      },
      "UsersResponseDataDto": {
        "type": "object",
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserResponseDto"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          },
          "totalUsers": {
            "type": "number",
            "example": 100
          }
        },
        "required": [
          "users",
          "pagination",
          "totalUsers"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "User name",
            "example": "John"
          },
          "surname": {
            "type": "string",
            "description": "User surname",
            "example": "Doe"
          },
          "dateOfBirth": {
            "format": "date-time",
            "type": "string",
            "description": "User date of birth",
            "example": "11.09.1990"
          },
          "email": {
            "type": "string",
            "description": "User email",
            "example": "johndoe@mail.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "P@ssw0rd123",
            "minLength": 8,
            "maxLength": 64
          },
          "oldPassword": {
            "type": "string",
            "description": "User old password",
            "example": "P@ssw0rd123"
          }
        }
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "User name",
            "example": "John"
          },
          "surname": {
            "type": "string",
            "description": "User surname",
            "example": "Doe"
          },
          "dateOfBirth": {
            "type": "string",
            "description": "User date of birth",
            "example": "1990-09-11"
          },
          "email": {
            "type": "string",
            "description": "User email",
            "example": "johndoe@mail.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "P@ssw0rd123",
            "minLength": 8,
            "maxLength": 64
          }
        },
        "required": [
          "name",
          "surname",
          "dateOfBirth",
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email",
            "example": "johndoe@mail.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "P@ssw0rd123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LogoutResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "User has been signed out"
          }
        },
        "required": [
          "message"
        ]
      },
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Post title"
          },
          "content": {
            "type": "string",
            "description": "Post content"
          },
          "image": {
            "type": "string",
            "description": "Post image url",
            "example": "https://firebasestorage.googleapis.com/exampleImage.jpg"
          }
        },
        "required": [
          "title",
          "content",
          "image"
        ]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Post title"
          },
          "content": {
            "type": "string",
            "description": "Post content"
          },
          "image": {
            "type": "string",
            "description": "Post image url",
            "example": "https://firebasestorage.googleapis.com/exampleImage.jpg"
          }
        }
      }
    }
  }
}